// apps/dokploy/server/utils/docker/types.ts
export type DefinitionsInclude =
	| string
	| {
			path?: StringOrList;
			env_file?: StringOrList;
			project_directory?: string;
	  };
export type StringOrList = string | ListOfStrings;
// ...
export type ListOrDict =
	| {
			/**
			 * This interface was referenced by `undefined`'s JSON-Schema definition
			 * via the `patternProperty` ".+".
			 */
			[k: string]: string | number | boolean | null;
	  }
	| string[];
// ...
export type DefinitionsNetwork = {
	name?: string;
	driver?: string;
	driver_opts?: {
		/**
		 * This interface was referenced by `undefined`'s JSON-Schema definition
		 * via the `patternProperty` "^.+$".
		 */
		[k: string]: string | number;
	};
	ipam?: {
		driver?: string;
		config?: {
			subnet?: string;
			ip_range?: string;
			gateway?: string;
			aux_addresses?: {
				/**
				 * This interface was referenced by `undefined`'s JSON-Schema definition
				 * via the `patternProperty` "^.+$".
				 */
				[k: string]: string;
			};
			/**
			 * This interface was referenced by `undefined`'s JSON-Schema definition
			 * via the `patternProperty` "^x-".
			 */
			[k: string]: unknown;
		}[];
		options?: {
			/**
			 * This interface was referenced by `undefined`'s JSON-Schema definition
			 * via the `patternProperty` "^.+$".
			 */
			[k: string]: string;
		};
		/**
		 * This interface was referenced by `undefined`'s JSON-Schema definition
		 * via the `patternProperty` "^x-".
		 */
		[k: string]: unknown;
	};
	external?:
		| boolean
		| {
				name?: string;
				/**
				 * This interface was referenced by `undefined`'s JSON-Schema definition
				 * via the `patternProperty` "^x-".
				 */
				[k: string]: unknown;
		  };
	internal?: boolean;
	enable_ipv6?: boolean;
	attachable?: boolean;
	labels?: ListOrDict;
	/**
	 * This interface was referenced by `undefined`'s JSON-Schema definition
	 * via the `patternProperty` "^x-".
	 *
	 * This interface was referenced by `undefined`'s JSON-Schema definition
	 * via the `patternProperty` "^x-".
	 */
	[k: string]: unknown;
} & Network;
// ...
export type DefinitionsVolume = {
	name?: string;
	driver?: string;
	driver_opts?: {
		/**
		 * This interface was referenced by `undefined`'s JSON-Schema definition
		 * via the `patternProperty` "^.+$".
		 */
		[k: string]: string | number;
	};
	external?:
		| boolean
		| {
				name?: string;
				/**
				 * This interface was referenced by `undefined`'s JSON-Schema definition
				 * via the `patternProperty` "^x-".
				 */
				[k: string]: unknown;
		  };
	labels?: ListOrDict;
	/**
	 * This interface was referenced by `undefined`'s JSON-Schema definition
	 * via the `patternProperty` "^x-".
	 *
	 * This interface was referenced by `undefined`'s JSON-Schema definition
	 * via the `patternProperty` "^x-".
	 */
	[k: string]: unknown;
} & Volume;
// ...
export interface ComposeSpecification {
	/**
	 * declared for backward compatibility, ignored.
	 */
	version?: string;
	/**
	 * define the Compose project name, until user defines one explicitly.
	 */
	name?: string;
	/**
	 * compose sub-projects to be included.
	 */
	include?: DefinitionsInclude[];
	services?: PropertiesServices;
	networks?: PropertiesNetworks;
	volumes?: PropertiesVolumes;
	secrets?: PropertiesSecrets;
	configs?: PropertiesConfigs;
	/**
	 * This interface was referenced by `ComposeSpecification`'s JSON-Schema definition
	 * via the `patternProperty` "^x-".
	 */
	[k: string]: unknown;
}
export interface PropertiesServices {
	[k: string]: DefinitionsService;
}
// ...
export interface PropertiesNetworks {
	[k: string]: DefinitionsNetwork;
}
export interface PropertiesVolumes {
	[k: string]: DefinitionsVolume;
}
export interface PropertiesSecrets {
	[k: string]: DefinitionsSecret;
}
/**
 * This interface was referenced by `PropertiesSecrets`'s JSON-Schema definition
 * via the `patternProperty` "^[a-zA-Z0-9._-]+$".
 */
export interface DefinitionsSecret {
	name?: string;
	environment?: string;
	file?: string;
	external?:
		| boolean
		| {
				name?: string;
				[k: string]: unknown;
		  };
	labels?: ListOrDict;
	driver?: string;
	driver_opts?: {
		/**
		 * This interface was referenced by `undefined`'s JSON-Schema definition
		 * via the `patternProperty` "^.+$".
		 */
		[k: string]: string | number;
	};
	template_driver?: string;
	/**
	 * This interface was referenced by `DefinitionsSecret`'s JSON-Schema definition
	 * via the `patternProperty` "^x-".
	 */
	[k: string]: unknown;
}
export interface PropertiesConfigs {
	[k: string]: DefinitionsConfig;
}
/**
 * This interface was referenced by `PropertiesConfigs`'s JSON-Schema definition
 * via the `patternProperty` "^[a-zA-Z0-9._-]+$".
 */
export interface DefinitionsConfig {
	name?: string;
	content?: string;
	environment?: string;
	file?: string;
	external?:
		| boolean
		| {
				name?: string;
				[k: string]: unknown;
		  };
	labels?: ListOrDict;
	template_driver?: string;
	/**
	 * This interface was referenced by `DefinitionsConfig`'s JSON-Schema definition
	 * via the `patternProperty` "^x-".
	 */
	[k: string]: unknown;
}